feat: make module federation final works
