+++ projects/services/kubestro-core/frontend/src/features/dashboard/_routes.ts
@@ -3,6 +3,7 @@ import homePage from './home/home'
 import notFoundPage from './not-found/not-found'
 import { settingsRoutes } from './settings/_routes'
 import { gameManagersRoute } from './game-managers/_routes'
+import { ErrorBoundary } from './_error-boundaries'
 
 export const dashboardRoutes: RouteObject[] = [
   homePage,
@@ -14,6 +15,11 @@ export const dashboardRoutes: RouteObject[] = [
     path: 'settings',
     lazy: async () => await import('./settings/_layout').then(m => m.default),
     children: settingsRoutes
+  },
+  {
+    id: 'modules',
+    children: [],
+    ErrorBoundary
   },
   {
     path: '*',
