+++ projects/services/module-federation-test/src/main.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
-import App from './app.tsx'
+import { App } from './app.tsx'
 
 const root = document.getElementById('root')
 
@@ -10,6 +10,6 @@ if (!root) {
 
 ReactDOM.createRoot(root).render(
   <React.StrictMode>
-    <App user="demo" />
+    <App />
   </React.StrictMode>
 )
